move ti.lib to ti._lib
