+++ MANIFEST.in
@@ -5,10 +5,10 @@ include *.cfg
 include python/taichi/*.md
 include python/taichi/assets/*
 recursive-include python/taichi/examples *.py
-include python/taichi/lib/core/*.so
-include python/taichi/lib/core/*.pyd
-include python/taichi/lib/runtime/*.bc
-include python/taichi/lib/runtime/*.dylib
+include python/taichi/_lib/core/*.so
+include python/taichi/_lib/core/*.pyd
+include python/taichi/_lib/runtime/*.bc
+include python/taichi/_lib/runtime/*.dylib
 include python/taichi/shaders/*.spv
 include python/taichi/shaders/*.vert
 include python/taichi/shaders/*.frag
