+++ python/.gitignore
@@ -1,6 +1,6 @@
-taichi/lib/runtime
-taichi/lib/core/*.so
-taichi/lib/core/*.pyd
+taichi/_lib/runtime
+taichi/_lib/core/*.so
+taichi/_lib/core/*.pyd
 taichi.egg-info
 taichi/include
 taichi/assets
