+++ python/taichi/_testing.py
@@ -6,7 +6,7 @@ from tempfile import mkstemp
 
 from taichi.lang import (cc, cpu, cuda, gpu, is_arch_supported, metal, opengl,
                          vulkan)
-from taichi.lib.core import ti_core as _ti_core
+from taichi._lib.core import ti_core as _ti_core
 
 import taichi as ti
