+++ python/taichi/lang/__init__.py
@@ -50,8 +50,8 @@ from taichi.lang.type_factory_impl import type_factory
 from taichi.lang.util import (cook_dtype, has_clangpp, has_pytorch,
                               is_taichi_class, python_scope, taichi_scope,
                               to_numpy_type, to_pytorch_type, to_taichi_type)
-from taichi.lib.core import ti_core as _ti_core
-from taichi.lib.core.util import locale_encode
+from taichi._lib.core import ti_core as _ti_core
+from taichi._lib.core.util import locale_encode
 from taichi.profiler import KernelProfiler, get_default_kernel_profiler
 from taichi.profiler.kernelmetrics import (CuptiMetric, default_cupti_metrics,
                                            get_predefined_cupti_metrics)
