+++ python/taichi/lang/field.py
@@ -1,6 +1,6 @@
 import taichi.lang
 from taichi.lang.util import python_scope, to_numpy_type, to_pytorch_type
-from taichi.lib.core import ti_core as _ti_core
+from taichi._lib.core import ti_core as _ti_core
 
 import taichi as ti
