+++ python/taichi/lang/matrix.py
@@ -13,7 +13,7 @@ from taichi.lang.exception import TaichiSyntaxError
 from taichi.lang.field import Field, ScalarField, SNodeHostAccess
 from taichi.lang.util import (cook_dtype, in_python_scope, python_scope,
                               taichi_scope, to_numpy_type, to_pytorch_type)
-from taichi.lib.core import ti_core
+from taichi._lib.core import ti_core
 from taichi.tools.util import deprecated, warning
 from taichi.types import CompoundType
