+++ python/taichi/lang/mesh.py
@@ -8,7 +8,7 @@ from taichi.lang.field import Field, ScalarField
 from taichi.lang.matrix import MatrixField, _IntermediateMatrix
 from taichi.lang.struct import StructField
 from taichi.lang.util import python_scope
-from taichi.lib.core import ti_core as _ti_core
+from taichi._lib.core import ti_core as _ti_core
 from taichi.types import CompoundType
 
 import taichi as ti
