+++ python/taichi/lang/type_factory_impl.py
@@ -1,5 +1,5 @@
 from taichi.lang import impl
-from taichi.lib.core import ti_core as _ti_core
+from taichi._lib.core import ti_core as _ti_core
 
 
 class TypeFactory:
