+++ python/taichi/lang/util.py
@@ -3,7 +3,7 @@ import os
 
 import numpy as np
 from taichi.lang import impl
-from taichi.lib.core import ti_core as _ti_core
+from taichi._lib.core import ti_core as _ti_core
 
 import taichi as ti
