+++ python/taichi/profiler/kernelprofiler.py
@@ -1,7 +1,7 @@
 from contextlib import contextmanager
 
 from taichi.lang import impl
-from taichi.lib.core import ti_core as _ti_core
+from taichi._lib.core import ti_core as _ti_core
 from taichi.profiler.kernelmetrics import default_cupti_metrics
 
 import taichi as ti
