+++ python/taichi/tools/image.py
@@ -1,7 +1,7 @@
 from io import BytesIO
 
 import numpy as np
-from taichi.lib.core import ti_core as _ti_core
+from taichi._lib.core import ti_core as _ti_core
 
 import taichi as ti
