+++ python/taichi/ui/utils.py
@@ -2,7 +2,7 @@ from math import acos, asin, cos, sin
 
 from taichi.lang.impl import default_cfg
 from taichi.lang.matrix import Vector
-from taichi.lib.core import ti_core as _ti_core
+from taichi._lib.core import ti_core as _ti_core
 
 
 def get_field_info(field):
